requirements.txt
telethon # ==1.28.5
requests # ==2.28.1
Pillow # ==9.4.0
beautifulsoup4 # ==4.11.1
python-dotenv # ==0.21.0
